Update README.md
This code performs a number of spelling changes in a given order. It has been conceived to update the spelling of Old Souletin texts (Eastern Basque) into present-day standard Basque spelling. I have used a function for case sensitive replacement, and a series of replacements by means of REGEX. Additionally, the project includes a function to count the changes operated by each rule. The resulting text would need of revision by a philologist, because variation in spelling includes exceptions to the rules. Nevertheless, I think the code is helpful to automatize the task of modernizing spelling.